Check for the existence of `kano-draw` progress within the project tracking, before checking for specific group
For users that have just signed up, or have no progress in Make Art, the `kano-draw` progress object is empty, so the app would throw an uncaught error  when trying to find sub-keys within the object. Uncaught errors were preventing the project listings from displaying.

Now the app checks for the `group` key before accessing sub-keys to avoid the error, and the project listings should not be interrupted.

This should fix this bug: https://trello.com/c/5x9uPif1/1185-new-users-who-complete-a-challenge-and-come-back-to-projects-have-missing-info