Add kw-auth-modal and kw-blank-project-modal as lazy dependencies to kw-app
This should resolve the problem reported in the Trello card below of the login modal on `/start` not firing. Suspected issue for this was that the `shards` task was not detecting these as dependencies of `kw-app` and therefore not bundling them into the build.

This highlighted a problem with the `shards` task, which is that it was trying to include depdencies in commented out links. A quick fix to this is to introduce a regex to detect for commented sections in the bower component html files, and strip these out, before running pattern checks for import links. This doesn't seem to add considerably to the build time. In future, it looks like we might switch to something like `parse5` for html parsing, so this method is likely to change. Hopefully this will fix the bug in the meantime!

Trello card: https://trello.com/c/iwzExmhA/1672-kano-world-log-in-on-start-is-not-working